+++ lib/pages/home.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'navigationBar.dart';
+
 class HomePage  extends StatefulWidget {
   const HomePage ({Key? key}) : super(key: key);
 
@@ -229,22 +231,6 @@ class _HomePageState extends State<HomePage> {
         ),
       ],
     );
-  }
-
-// 각각의 버튼이 호출할 함수들
-  void onRankingPressed() {
-    // "랭킹" 버튼을 눌렀을 때 실행할 코드
-    print("랭킹 버튼이 눌렸습니다");
-  }
-
-  void onHomePressed() {
-    // "홈" 버튼을 눌렀을 때 실행할 코드
-    print("홈 버튼이 눌렸습니다");
-  }
-
-  void onProfilePressed() {
-    // "마이" 버튼을 눌렀을 때 실행할 코드
-    print("마이 버튼이 눌렸습니다");
   }
