+++ lib/pages/navigationBar.dart
@@ -1,0 +1,15 @@
+// 각각의 버튼이 호출할 함수들
+void onRankingPressed() {
+  // "랭킹" 버튼을 눌렀을 때 실행할 코드
+  print("랭킹 버튼이 눌렸습니다");
+}
+
+void onHomePressed() {
+  // "홈" 버튼을 눌렀을 때 실행할 코드
+  print("홈 버튼이 눌렸습니다");
+}
+
+void onProfilePressed() {
+  // "마이" 버튼을 눌렀을 때 실행할 코드
+  print("마이 버튼이 눌렸습니다");
+}
